refactor(watch): extract useChannelLiveDiff (WS diff + viewer debounce)
